🐛(fix): casting int in epic_id
"dw was receiving the wrong type

+++ src/main/java/com/manolito/dashflow/transformer/TaigaTransformer.java
@@ -28,7 +28,7 @@ public class TaigaTransformer {
                 col("id").as("original_id"),
                 lit(TOOL_ID).as("tool_id"),
                 col("project").as("project_id"),
-                col("epics").as("epic_id"),
+                col("epics").cast("int").as("epic_id"),
                 col("subject").as("story_name"),
                 col("is_closed").as("is_finished")
         );
